+++ src/main/kotlin/io/github/t45k/clione/controller/cloneDetector/NiCadController.kt
@@ -2,9 +2,9 @@ package io.github.t45k.clione.controller.cloneDetector
 
 import com.github.kusumotolab.sdl4j.util.CommandLine
 import com.google.common.annotations.VisibleForTesting
-import io.github.t45k.clione.core.Granularity
-import io.github.t45k.clione.core.Language
-import io.github.t45k.clione.core.RunningConfig
+import io.github.t45k.clione.core.config.Granularity
+import io.github.t45k.clione.core.config.Language
+import io.github.t45k.clione.core.config.RunningConfig
 import io.github.t45k.clione.entity.CloneInstance
 import io.github.t45k.clione.entity.CloneSets
 import io.github.t45k.clione.entity.CloneStatus
