+++ src/main/kotlin/io/github/t45k/clione/controller/cloneDetector/cloneCandidate/CloneCandidateExtractor.kt
@@ -8,13 +8,13 @@ import io.github.t45k.clione.controller.cloneDetector.cloneCandidate.kotlin.Kotl
 import io.github.t45k.clione.controller.cloneDetector.cloneCandidate.kotlin.KotlinMethodExtractor
 import io.github.t45k.clione.controller.cloneDetector.cloneCandidate.python3.Python3BlockExtractor
 import io.github.t45k.clione.controller.cloneDetector.cloneCandidate.python3.Python3FunctionExtractor
-import io.github.t45k.clione.core.Granularity.BLOCK
-import io.github.t45k.clione.core.Granularity.METHOD
-import io.github.t45k.clione.core.Language.CPP
-import io.github.t45k.clione.core.Language.JAVA
-import io.github.t45k.clione.core.Language.KOTLIN
-import io.github.t45k.clione.core.Language.PYTHON
-import io.github.t45k.clione.core.RunningConfig
+import io.github.t45k.clione.core.config.Granularity.BLOCK
+import io.github.t45k.clione.core.config.Granularity.METHOD
+import io.github.t45k.clione.core.config.Language.CPP
+import io.github.t45k.clione.core.config.Language.JAVA
+import io.github.t45k.clione.core.config.Language.KOTLIN
+import io.github.t45k.clione.core.config.Language.PYTHON
+import io.github.t45k.clione.core.config.RunningConfig
 import io.github.t45k.clione.entity.CloneStatus
 import io.github.t45k.clione.entity.InvalidConfigSpecifiedException
 import java.nio.file.Path
